+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 #### Fixed
 - [#121]: ContentTypeEngines were not properly initialized during the discovery/registration process 
 - Fixed multiple regex parameter tokenization
+- [#132]: Prevent web server instantiation
 
 #### Changed
 - [#117]: Rework [pippo-spring], [pippo-guice]
@@ -115,6 +116,7 @@ Initial release.
 [0.3.0]: https://github.com/decebals/pippo/compare/pippo-parent-0.2.0...pippo-parent-0.3.0
 [0.2.0]: https://github.com/decebals/pippo/compare/pippo-parent-0.1.0...pippo-parent-0.2.0
 
+[#132]: https://github.com/decebals/pippo/issues/132
 [#134]: https://github.com/decebals/pippo/issues/134
 [#131]: https://github.com/decebals/pippo/issues/131
 [#130]: https://github.com/decebals/pippo/issues/130