+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
+#### Fixed
+#### Changed
+#### Added
+#### Removed
+
+### [0.5.0] - 2015-06-03
 
 #### Fixed
 - [#121]: ContentTypeEngines were not properly initialized during the discovery/registration process 
@@ -109,15 +115,16 @@ Initial release.
 - Added [pippo-freemarker]
 - Added [pippo-jetty]
 
-[unreleased]: https://github.com/decebals/pippo/compare/release-0.4.2...HEAD
+[unreleased]: https://github.com/decebals/pippo/compare/release-0.5.0...HEAD
+[0.5.0]: https://github.com/decebals/pippo/compare/release-0.4.2...release-0.5.0
 [0.4.2]: https://github.com/decebals/pippo/compare/release-0.4.1...release-0.4.2
 [0.4.1]: https://github.com/decebals/pippo/compare/pippo-parent-0.4.0...release-0.4.1
 [0.4.0]: https://github.com/decebals/pippo/compare/pippo-parent-0.3.0...pippo-parent-0.4.0
 [0.3.0]: https://github.com/decebals/pippo/compare/pippo-parent-0.2.0...pippo-parent-0.3.0
 [0.2.0]: https://github.com/decebals/pippo/compare/pippo-parent-0.1.0...pippo-parent-0.2.0
 
-[#132]: https://github.com/decebals/pippo/issues/132
 [#134]: https://github.com/decebals/pippo/issues/134
+[#132]: https://github.com/decebals/pippo/issues/132
 [#131]: https://github.com/decebals/pippo/issues/131
 [#130]: https://github.com/decebals/pippo/issues/130
 [#129]: https://github.com/decebals/pippo/issues/129